@@ -1,22 +1,19 @@
-import Vue from 'vue'
-
-// import { BootstrapVue } from 'bootstrap-vue'
-import { LayoutPlugin, NavbarPlugin, ButtonPlugin, LinkPlugin, ModalPlugin } from 'bootstrap-vue'
-import 'bootstrap/dist/css/bootstrap.css'
-import 'bootstrap-vue/dist/bootstrap-vue.css'
-
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
-import store from './store'
 
-Vue.use(LayoutPlugin)
-Vue.use(NavbarPlugin)
-Vue.use(ButtonPlugin)
-Vue.use(LinkPlugin)
-Vue.use(ModalPlugin)
+import { createStore } from 'vuex'
+import { store } from './store'
 
-new Vue({
-  router,
-  store,
-  render: h => h(App)
-}).$mount('#app')
+// import { BootstrapVue } from 'bootstrap-vue-next'
+import 'bootstrap/dist/css/bootstrap.css'
+import 'bootstrap-vue-next/dist/bootstrap-vue-next.css'
+// import './assets/main.css'
+
+const storeInstance = createStore(store)
+const app = createApp(App)
+
+app.use(router)
+app.use(storeInstance)
+
+app.mount('#app')
