@@ -1,6 +1,10 @@
+<script setup>
+import { RouterView } from 'vue-router'
+</script>
+
 <template>
   <div id="app">
-    <router-view />
+    <RouterView />
   </div>
 </template>
 
