@@ -9,12 +9,12 @@
 </template>
 
 <script>
-import { BContainer } from 'bootstrap-vue'
+import { BContainer } from 'bootstrap-vue-next'
 import TrackManagerNavBar from '@/components/TrackManagerNavBar.vue'
 export default {
   name: 'TrackOverviewPage',
   components: {
-    TrackOverview: () => import(/* webpackChunkName: "TrackOverView" */ '@/components/TrackOverview.vue'),
+    TrackOverview: () => import('@/components/TrackOverview.vue'),
     TrackManagerNavBar,
     BContainer
   },
